+++ src/index.js
@@ -1,9 +1,15 @@
-const Tx = require('ethereumjs-tx');
-const ethUtil = require('ethjs-util');
 const HTTPProvider = require('ethjs-provider-http');
 
 module.exports = SignerProvider;
 
+/**
+ * Signer provider constructor
+ *
+ * @method SignerProvider
+ * @param {String} path the input data payload
+ * @param {Object} options the send async callback
+ * @returns {Object} provider instance
+ */
 function SignerProvider(path, options) {
   if (!(this instanceof SignerProvider)) { throw new Error('[ethjs-provider-signer] the SignerProvider instance requires the "new" flag in order to function normally (e.g. `const eth = new Eth(new SignerProvider(...));`).'); }
   if (typeof options !== 'object') { throw new Error(`[ethjs-provider-signer] the SignerProvider requires an options object be provided with the 'privateKey' property specified, you provided type ${typeof options}.`); }
@@ -15,32 +21,14 @@ function SignerProvider(path, options) {
   self.provider = new HTTPProvider(path, self.timeout);
 }
 
-// fix ethereumjs-tx rawTx object
-function shimEthereumJSTxObject(rawTx) {
-  const rawTxMutation = Object.assign({}, rawTx);
-
-  // fix rawTx gaslimit
-  if (typeof rawTxMutation.gas === 'string') {
-    rawTxMutation.gasLimit = rawTxMutation.gas;
-    delete rawTxMutation.gas;
-  }
-
-  // return new mutated raw tx object
-  return rawTxMutation;
-}
-
-SignerProvider.sign = function signTx(rawTx, privateKey) {
-  // format raw tx data
-  const shimmedRawTx = Object.assign({}, shimEthereumJSTxObject(rawTx));
-
-  // sign tx object, serilize, convert to hex, privateKey not stored in temp data
-  const signedHexPayload = new Tx(shimmedRawTx);
-  signedHexPayload.sign(new Buffer(ethUtil.stripHexPrefix(privateKey), 'hex'));
-
-  // ensure hex is prefixed
-  return `0x${signedHexPayload.serialize().toString('hex')}`;
-};
-
+/**
+ * Send async override
+ *
+ * @method sendAsync
+ * @param {payload} payload the input data payload
+ * @param {Function} callback the send async callback
+ * @callback {Object} output the XMLHttpRequest payload
+ */
 SignerProvider.prototype.sendAsync = function (payload, callback) { // eslint-disable-line
   const self = this;
   if (payload.method === 'eth_sendTransaction') {