+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 0.1.2 -- remove deps
+
+1. Remove `ethereumjs-tx` and `ethereumjs-util`
+2. Uses `ethjs-signer` for signing now, no need to include it in the module
+
+# 0.1.1 -- config fixes
+
+1. webpack config updates
+2. build config updates
+
 # 0.0.1 -- ethjs-provider-signer
 
 1. Basic testing